# backend/swagger.yaml
openapi: 3.0.1
info:
  title: Mini LMS API
  version: "1.0.0"
  description: |
    RESTful API for the Mini Learning Management System platform.  
    Includes authentication, courses, enrollments, quizzes, attempts, certificates, and admin dashboard endpoints.
servers:
  - url: http://localhost:5000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin, staff]
    Course:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        instructor:
          $ref: '#/components/schemas/User'
        studentsEnrolled:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
security:
  - bearerAuth: []
paths:
...
  /courses:
    get:
      tags: [Courses]
      summary: Get all courses
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
                  pagination:
                    type: object
    post:
      tags: [Courses]
      summary: Create a new course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                thumbnail:
                  type: string
                  format: binary
            required: [title, description]
      responses:
        '201':
          description: Course created
  /courses/{courseId}:
    get:
      tags: [Courses]
      summary: Get course by ID
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course detail
        '404':
          description: Course not found
    put:
      tags: [Courses]
      summary: Update course
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                thumbnail:
                  type: string
                  format: binary
      responses:
        '200':
          description: Course updated
    delete:
      tags: [Courses]
      summary: Delete course
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...
  /admin/courses:
    get:
      tags: [Admin]
      summary: List courses with admin stats
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
      responses:
        '200':
          description: Admin course list
...